+++ app/Main.hs
@@ -76,155 +76,8 @@ import System.IO
 import System.Process
 import Data.String (IsString)
 
-data HostPort = HostPort {_host :: Text, _port :: Int}
-  deriving (Show, Eq)
-
-makeLenses ''HostPort
-
-theMap :: AttrMap
-theMap =
-  attrMap
-    V.defAttr
-    [ (E.editAttr, V.white `on` V.black),
-      (E.editFocusedAttr, V.black `on` V.yellow),
-      (invalidFormInputAttr, V.white `on` V.red),
-      (focusedFormInputAttr, V.black `on` V.yellow),
-      ("header", V.black `on` V.white),
-      (L.listAttr, V.white `on` V.black),
-      (L.listSelectedAttr, V.white `on` V.blue)
-    ]
-
-data VibeMenuName
-  = MessageLog
-  | DeviceMenu
-  | HostField
-  | PortField
-  deriving (Eq, Ord, Show)
-
--- Commands from the UI thread to the background thread
-data ButtplugCommand
-  = CmdStopAll
-  | CmdVibrate Word Double
-  deriving (Show)
-
--- TODO give BPWS.Connector a `Show` instance so we can derive `Show` here
-data Command
-  = BPCommand ButtplugCommand
-  | CmdConnect BPWS.Connector
-
--- events from bg thread to UI thread
-data VibeMenuEvent = EvConnected | BPEvent BPSessionEvent
-
--- events from buttplug server
-data BPSessionEvent
-  = ReceivedMessage Message
-  | ReceivedDeviceList [Device]
-  | EvDeviceAdded Device
-  | EvDeviceRemoved Word
-
-data MainScreenState = MainScreenState
-  { _messageLog :: L.List VibeMenuName Message,
-    _devices :: L.List VibeMenuName Device
-  }
-
-type ConnectForm = Form HostPort VibeMenuEvent VibeMenuName
-
-mkConnectForm :: HostPort -> ConnectForm
-mkConnectForm =
-  let label s w =
-        padBottom (Pad 1) $
-          vLimit 1 (hLimit 15 $ str s <+> fill ' ') <+> w
-   in newForm
-        [ label "Host"
-            @@= editTextField host HostField (Just 1),
-          label "Port"
-            @@= editShowableField port PortField
-        ]
-
-makeLenses ''MainScreenState
-
-data ScreenState
-  = ConnectScreen ConnectForm
-  | ConnectingScreen
-  | MainScreen MainScreenState
-
-data AppState = AppState {_cmdChan :: BChan Command, _screenState :: ScreenState}
-
-makeLenses ''AppState
-
-mainScreenState :: Lens' ScreenState MainScreenState
-mainScreenState = lens get set
-  where
-    get = \case
-      MainScreen mss -> mss
-      _ -> error "BUG: mainScreenState: tried to get MainScreenState from incorrect constructor"
-
-    set appState mss = case appState of
-      MainScreen _ -> MainScreen mss
-      _ ->
-        error
-          "BUG: mainScreenState: tried to write MainScreenState to incorrect constructor field"
-
-connectScreenState :: Lens' ScreenState ConnectForm
-connectScreenState = lens get set
-  where
-    get = \case
-      ConnectScreen connectForm -> connectForm
-      _ -> error "BUG: connectScreenState: tried to get ConnectForm from incorrect constructor"
-
-    set appState connectForm = case appState of
-      ConnectScreen _ -> ConnectScreen connectForm
-      _ ->
-        error
-          "BUG: connectScreenState: tried to write ConnectForm to incorrect constructor field"
-
-drawVibeMenu s = case s ^. screenState of
-  ConnectScreen _ -> drawConnectScreen s
-  ConnectingScreen -> drawConnectingScreen s
-  MainScreen _ -> drawMainScreen s
-
-drawConnectScreen :: AppState -> [Widget VibeMenuName]
-drawConnectScreen s = [C.vCenter $ C.hCenter form]
-  where
-    f = s ^. screenState . connectScreenState
-    form = B.border $ padTop (Pad 1) $ hLimit 40 $ renderForm f
-
-drawConnectingScreen :: AppState -> [Widget VibeMenuName]
-drawConnectingScreen s = [txt "Connecting..."]
-
-drawMainScreen :: AppState -> [Widget VibeMenuName]
-drawMainScreen appState = [ui]
-  where
-    s = appState ^. screenState . mainScreenState
-    header = withAttr "header" . txtWrap
-    title = padBottom (Pad 1) $ header "VibeMenu"
-    deviceMenu =
-      header "Connected Devices"
-        <=> padBottom (Pad 1) (L.renderList listDrawDevice True $ s ^. devices)
-    receivedMsgLog =
-      header "Message log"
-        <=> padBottom (Pad 1) (L.renderList listDrawElement False $ s ^. messageLog)
-
-    ui =
-      title
-        <=> deviceMenu
-        <=> receivedMsgLog
-
-listDrawDevice :: Bool -> Device -> Widget VibeMenuName
-listDrawDevice sel Device {..}
-  | sel = withAttr L.listSelectedAttr $ txt label
-  | otherwise = txt label
-  where
-    label :: T.Text
-    label = T.pack (show deviceIndex) <> " " <> deviceName
-
-listDrawElement :: (Show e) => Bool -> e -> Widget VibeMenuName
-listDrawElement sel a =
-  let selStr s =
-        if sel
-          then withAttr L.listSelectedAttr $ str s
-          else str s
-   in (selStr $ show a)
+import Types
+import View
 
 vibeMenu ::
   (AppState -> EventM VibeMenuName AppState) ->
@@ -348,17 +201,6 @@ mainScreenHandleEvent s = \case
     outOfTen = \case
       0 -> 1
       n -> fromIntegral n / 10
-
-listAppend :: e -> L.List n e -> L.List n e
-listAppend elt l =
-  let n = Vec.length $ L.listElements l
-   in L.listInsert n elt l
-
-deleteDeviceByIndex :: Word -> L.List n Device -> L.List n Device
-deleteDeviceByIndex devIx l =
-  case Vec.findIndex ((devIx ==) . deviceIndex) (L.listElements l) of
-    Just listIndex -> L.listRemove listIndex l
-    Nothing -> l
 
 hostPortToConnector (HostPort host port) = BPWS.Connector (T.unpack host) port
